+++ native~/Runtime/src/TextureLoader.cpp
@@ -40,9 +40,10 @@ TextureLoader::loadTexture(const CesiumGltf::ImageCesium& image) {
   assert(textureLength >= image.pixelData.size());
 
   if (image.mipPositions.empty()) {
-    // No mipmaps, copy the whole thing and then let Unity generate mipmaps.
+    // No mipmaps, copy the whole thing and then let Unity generate mipmaps on a
+    // worker thread.
     std::memcpy(pixels, image.pixelData.data(), image.pixelData.size());
-    result.Apply(true, true);
+    result.Apply(false, true);
   } else {
     // Copy the mipmaps explicitly.
     std::uint8_t* pWritePosition = pixels;