+++ native~/Runtime/src/UnityPrepareRendererResources.cpp
@@ -13,6 +13,7 @@
 #include <CesiumGltf/ExtensionKhrMaterialsUnlit.h>
 #include <CesiumGltf/ExtensionMeshPrimitiveExtFeatureMetadata.h>
 #include <CesiumGltf/ExtensionModelExtFeatureMetadata.h>
+#include <CesiumGltfReader/GltfReader.h>
 #include <CesiumShaderProperties.h>
 #include <CesiumUtility/ScopeGuard.h>
 
@@ -240,6 +241,40 @@ bool validateVertexColors(
   }
 
   return true;
+}
+
+void generateMipMaps(
+    Model* pModel,
+    const std::optional<TextureInfo>& textureInfo) {
+  if (textureInfo) {
+    Texture* pTexture = Model::getSafe(&pModel->textures, textureInfo->index);
+    if (pTexture) {
+      Image* pImage = Model::getSafe(&pModel->images, pTexture->source);
+      if (pImage) {
+        CesiumGltfReader::GltfReader::generateMipMaps(pImage->cesium);
+      }
+    }
+  }
+}
+
+void generateMipMapsForPrimitive(
+    Model* pModel,
+    const MeshPrimitive& primitive) {
+  const Material* pMaterial =
+      Model::getSafe(&pModel->materials, primitive.material);
+  if (pMaterial) {
+    if (pMaterial->pbrMetallicRoughness) {
+      generateMipMaps(
+          pModel,
+          pMaterial->pbrMetallicRoughness->baseColorTexture);
+      generateMipMaps(
+          pModel,
+          pMaterial->pbrMetallicRoughness->metallicRoughnessTexture);
+    }
+    generateMipMaps(pModel, pMaterial->normalTexture);
+    generateMipMaps(pModel, pMaterial->occlusionTexture);
+    generateMipMaps(pModel, pMaterial->emissiveTexture);
+  }
 }
 
 template <typename TIndex, class TIndexAccessor>
@@ -599,10 +634,12 @@ int32_t countPrimitives(const CesiumGltf::Model& model) {
   return numberOfPrimitives;
 }
 
+
+
 void populateMeshDataArray(
     MeshDataResult& meshDataResult,
-    const TileLoadResult& tileLoadResult) {
-  const CesiumGltf::Model* pModel =
+    TileLoadResult& tileLoadResult) {
+  CesiumGltf::Model* pModel =
       std::get_if<CesiumGltf::Model>(&tileLoadResult.contentKind);
   if (!pModel)
     return;
@@ -613,7 +650,7 @@ void populateMeshDataArray(
 
   pModel->forEachPrimitiveInScene(
       -1,
-      [&meshDataResult, &meshDataInstance](
+      [&meshDataResult, &meshDataInstance, pModel](
           const Model& gltf,
           const Node& node,
           const Mesh& mesh,
@@ -638,6 +675,9 @@ void populateMeshDataArray(
           // TODO: report invalid accessor
           return;
         }
+
+        generateMipMapsForPrimitive(pModel, primitive);
+
         if (primitive.indices < 0 ||
             primitive.indices >= gltf.accessors.size()) {
           int32_t indexCount = static_cast<int32_t>(positionView.size());
@@ -1374,6 +1414,7 @@ void UnityPrepareRendererResources::free(
 void* UnityPrepareRendererResources::prepareRasterInLoadThread(
     CesiumGltf::ImageCesium& image,
     const std::any& rendererOptions) {
+  CesiumGltfReader::GltfReader::generateMipMaps(image);
   return nullptr;
 }